requirements.txt
# Core Flask Dependencies
Flask==2.3.3
Flask-Limiter==2.8.1
Flask-Cors==4.0.0
werkzeug==2.3.7

# Environment & Configuration
python-dotenv==1.0.0

# System Monitoring
psutil==5.9.6

# HTTP Client (Required for Google OAuth)
requests==2.31.0

# Note: No authlib needed - using native requests for OAuth
